requirements/common.txt
alembic 
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.8.1
click 
clikit==0.6.2
crashtest==0.3.1
cryptography==36.0.1
distlib==0.3.4
fastapi 
fastapi-jwt-auth
FastAPI-SQLAlchemy
flake8 
itsdangerous 
Jinja2 
opentelemetry-api 
opentelemetry-instrumentation 
opentelemetry-instrumentation-asgi 
opentelemetry-instrumentation-fastapi 
opentelemetry-instrumentation-redis 
opentelemetry-instrumentation-sqlalchemy 
opentelemetry-sdk 
opentelemetry-semantic-conventions
opentelemetry-util-http 
passlib 
Pillow 
poetry==1.1.13
poetry-core==1.0.8
prometheus-client 
psycopg2-binary 
ptyprocess==0.7.0
pycodestyle 
pycparser==2.21
pydantic 
PyJWT 
pylev==1.4.0
pyparsing==3.0.7
redis 
requests==2.27.1
requests-toolbelt==0.9.1
rq
six==1.16.0
SQLAlchemy
sqlalchemy-mptt 
starlette 
starlette-prometheus
tomlkit==0.10.0
typer
typing_extensions 
uvicorn
virtualenv==20.13.3
webencodings==0.5.1
wrapt 
zipp==3.7.0
